requirements.txt
pandas
requests
beautifulsoup4
sqlalchemy
psycopg2
python-dotenv
jupyter